+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased](https://github.com/sensu/sensu-go-has-contact-filter/tree/HEAD)
 [Full Changelog](https://github.com/sensu/sensu-go-has-contact-filter/compare/0.2.0...HEAD)
 
+- `has_contact` function now trims empty space around contact names, both when present in comma-separated label values, and when present in contact name provided as a function argument.
+
 ## [0.2.0](https://github.com/sensu/sensu-go-has-contact-filter/tree/0.2.0) (2019-06-26)
 [Full Changelog](https://github.com/sensu/sensu-go-has-contact-filter/compare/0.1.2...0.2.0)